ase20_supplementary/requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
numpy==1.18.2
protobuf==3.7.1
protobuf3==0.2.1
pycparser==2.19
pycparserext==2019.1
scipy==1.1.0
torch-cluster==1.4.5
torch-geometric==1.3.2
torch-scatter==1.3.2
torch-sparse==0.4.3
tqdm==4.36.1
transformers==2.2.2
Werkzeug==0.16.0
lmdb==0.98
